+++ NovaChat/app/src/main/java/com/novachat/ui/navigation/NavRoutes.kt
@@ -61,3 +61,6 @@ object NotificationProfilesRoute
 
 @Serializable
 object SmartSecureRoute
+
+@Serializable
+object TrustedSendersRoute